storeextramaterial update changes

+++ app/src/main/java/com/android/gandharvms/Inward_Tanker_Security/Inward_Tanker_Security.java
@@ -644,7 +644,7 @@ public class Inward_Tanker_Security extends AppCompatActivity implements View.On
                             // Create a new JSONObject for each material
                             JSONObject materialObject = new JSONObject();
                             materialObject.put("Material", dynamaterial);
-                            materialObject.put("Qty", Integer.parseInt(dynaqty));
+                            materialObject.put("Qty", Double.parseDouble(dynaqty));
                             materialObject.put("Qtyuom", dynaqtyuom);
 
                             // Add the material JSON object to the array
